undo changes in original sample app
