+++ sample/cf-nodejs-logging-support-sample/app.js
@@ -1,12 +1,7 @@
-process.env.TEST_SENSITIVE_DATA = false;
-// saves public key
-process.env.DYN_LOG_LEVEL_KEY = "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA2fzU8StO511QYoC+BZp4riR2eVQM8FPPB2mF4I78WBDzloAVTaz0Z7hkMog1rAy8+Xva+fLiMuxDmN7kQZKBc24O4VeKNjOt8ZtNhz3vlMTZrNQ7bi+j8TS8ycUgKqe4/hSmjJBfXoduZ8Ye90u8RRfPLzbuutctLfCnL/ZhEehqfilt1iQb/CRCEsJou5XahmvOO5Gt+9kTBmY+2rS/+HKKdAhI3OpxwvXXNi8m9LrdHosMD7fTUpLUgdcIp8k3ACp9wCIIxbv1ssDeWKy7bKePihTl7vJq6RkopS6GvhO6yiD1IAJF/iDOrwrJAWzanrtavUc1RJZvbOvD0DFFOwIDAQAB";
-
 const express = require("express");
-// var log = require("../../v2/cf-nodejs-logging-support/build/main/index");
-var log = require("../../index.js");
+var log = require("cf-nodejs-logging-support");
 const app = express();
-// log.addConfig(require("./config-test.json"));
+
 // roots the views directory to public
 app.set('views', 'public');
 
@@ -33,7 +28,8 @@ var stats = {
   pid: process.pid,
   platform: process.platform,
 };
-//log.info("Message logged in global context with custom fields", stats);
+log.info("Message logged in global context with custom fields", stats);
+
 // home route
 app.get("/", function (req, res) {
   res.send();
@@ -42,14 +38,7 @@ app.get("/", function (req, res) {
 // demonstrate log in global context
 // https://sap.github.io/cf-nodejs-logging-support/general-usage/logging-contexts#global-context
 app.get("/globalcontext", function (req, res) {
-  var startTime = performance.now();
-
-  for (let index = 0; index < 1000; index++) {
-    log.logMessage("info", "Message logged in global context");
-  }
-
-  var endTime = performance.now();
-  console.log(`Call to doSomething took ${endTime - startTime} milliseconds`)
+  log.info("Message logged in global context");
   res.send();
 });
 
@@ -57,24 +46,14 @@ app.get("/globalcontext", function (req, res) {
 // https://sap.github.io/cf-nodejs-logging-support/general-usage/logging-contexts#request-context
 app.get("/requestcontext", function (req, res) {
   var reqLogger = req.logger; // reqLogger logs in request context
-
-  var startTime = performance.now()
-
-  for (let index = 0; index < 1000; index++) {
-    reqLogger.logMessage("info", "Message logged in request context");
-  }
-
-  var endTime = performance.now()
-  console.log(`Call to doSomething took ${endTime - startTime} milliseconds`)
-
+  reqLogger.info("Message logged in request context");
   res.send();
 });
 
 // log message with some custom fields in global context 
 // https://sap.github.io/cf-nodejs-logging-support/general-usage/custom-fields
 app.get("/customfields", function (req, res) {
-  log.setCustomFieldsFormat("application-logging");
-  log.logMessage("info", "Message logged in global context with some custom fields", { "custom-field-a": "value-a", "custom-field-b": "value-b" });
+  log.info("Message logged in global context with some custom fields", { "custom-field-a": "value-a", "custom-field-b": "value-b" });
   res.send();
 });
 
@@ -85,7 +64,7 @@ app.get("/stacktrace", function (req, res) {
     alwaysError();
     res.send("request succesful");
   } catch (e) {
-    log.logMessage("error", "Error occurred", e)
+    log.error("Error occurred", e)
     res.status(500).send("error ocurred");
   }
 });
@@ -99,7 +78,7 @@ function alwaysError() {
 app.get("/childlogger", function (req, res) {
   var subLogger = log.createLogger({ "new-field": "value" });
   subLogger.setLoggingLevel("warn");
-  subLogger.logMessage("warn", "Message logged from child logger.");
+  subLogger.warn("Message logged from child logger.");
   res.send();
 });
 
@@ -109,7 +88,7 @@ app.get("/correlationandtenantid", function (req, res) {
   var reqLogger = req.logger; // reqLogger logs in request context
   var correlationId = reqLogger.getCorrelationId();
   var tenantId = reqLogger.getTenantId();
-  reqLogger.logMessage("info", "Correlation ID: %s Tenant ID: %s", correlationId, tenantId);
+  reqLogger.info("Correlation ID: %s Tenant ID: %s", correlationId, tenantId);
   res.send();
 });
 
@@ -117,5 +96,5 @@ app.get("/correlationandtenantid", function (req, res) {
 // binds the express module to 'app', set port and run server
 var port = Number(process.env.VCAP_APP_PORT || 8080);
 app.listen(port, function () {
-  log.logMessage("info", "listening on port: %d", port);
+  log.info("listening on port: %d", port);
 });
