+++ v2/cf-nodejs-logging-support/src/performance-test/cf-nodejs-logging-support-sample/app.js
@@ -1,10 +1,37 @@
+process.env.VCAP_SERVICES = "CF";
+process.env.VCAP_APPLICATION = JSON.stringify(
+  {
+    "cf_api": "test-value",
+    "limits": {
+      "fds": 32768
+    },
+    "application_name": "test-value",
+    "application_uris": [
+      "test-value"
+    ],
+    "name": "test-value",
+    "space_name": "dev",
+    "space_id": "test-value",
+    "organization_id": "test-value",
+    "organization_name": "test-value",
+    "uris": [
+      "test-value"
+    ],
+    "users": null,
+    "application_id": "test-value"
+  }
+);
+process.env.LOG_SENSITIVE_CONNECTION_DATA = false;
+process.env.LOG_REMOTE_USER = false;
+process.env.LOG_REFERER = false;
+
 process.env.TEST_SENSITIVE_DATA = false;
 // saves public key
 process.env.DYN_LOG_LEVEL_KEY = "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA2fzU8StO511QYoC+BZp4riR2eVQM8FPPB2mF4I78WBDzloAVTaz0Z7hkMog1rAy8+Xva+fLiMuxDmN7kQZKBc24O4VeKNjOt8ZtNhz3vlMTZrNQ7bi+j8TS8ycUgKqe4/hSmjJBfXoduZ8Ye90u8RRfPLzbuutctLfCnL/ZhEehqfilt1iQb/CRCEsJou5XahmvOO5Gt+9kTBmY+2rS/+HKKdAhI3OpxwvXXNi8m9LrdHosMD7fTUpLUgdcIp8k3ACp9wCIIxbv1ssDeWKy7bKePihTl7vJq6RkopS6GvhO6yiD1IAJF/iDOrwrJAWzanrtavUc1RJZvbOvD0DFFOwIDAQAB";
 
 const express = require("express");
-// var log = require("../../v2/cf-nodejs-logging-support/build/main/index");
-var log = require("../../index.js");
+// var log = require("cf-nodejs-logging-support");
+var log = require("../../../build/main/index");
 const app = express();
 // log.addConfig(require("./config-test.json"));
 // roots the views directory to public
@@ -42,14 +69,19 @@ app.get("/", function (req, res) {
 // demonstrate log in global context
 // https://sap.github.io/cf-nodejs-logging-support/general-usage/logging-contexts#global-context
 app.get("/globalcontext", function (req, res) {
-  var startTime = performance.now();
+  // var startTime = performance.now();
 
-  for (let index = 0; index < 1000; index++) {
+  for (let index = 0; index < 100000; index++) {
     log.logMessage("info", "Message logged in global context");
   }
+  // var endTime = performance.now();
+  // console.log(`Call to doSomething took ${endTime - startTime} milliseconds`)
+  res.send();
+});
 
-  var endTime = performance.now();
-  console.log(`Call to doSomething took ${endTime - startTime} milliseconds`)
+// demonstrate log in global context
+// https://sap.github.io/cf-nodejs-logging-support/general-usage/logging-contexts#global-context
+app.get("/testlognetwork", function (req, res) {
   res.send();
 });
 
@@ -58,14 +90,14 @@ app.get("/globalcontext", function (req, res) {
 app.get("/requestcontext", function (req, res) {
   var reqLogger = req.logger; // reqLogger logs in request context
 
-  var startTime = performance.now()
+  // var startTime = performance.now()
 
-  for (let index = 0; index < 1000; index++) {
+  for (let index = 0; index < 100000; index++) {
     reqLogger.logMessage("info", "Message logged in request context");
   }
 
-  var endTime = performance.now()
-  console.log(`Call to doSomething took ${endTime - startTime} milliseconds`)
+  // var endTime = performance.now()
+  // console.log(`Call to doSomething took ${endTime - startTime} milliseconds`)
 
   res.send();
 });
@@ -115,7 +147,7 @@ app.get("/correlationandtenantid", function (req, res) {
 
 
 // binds the express module to 'app', set port and run server
-var port = Number(process.env.VCAP_APP_PORT || 8080);
+var port = Number(process.env.VCAP_APP_PORT || 8081);
 app.listen(port, function () {
   log.logMessage("info", "listening on port: %d", port);
 });
